added product ID, color code in the product creation

+++ database/migrations/2025_05_12_094833_create_supply_orders_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('supply_orders')) {
+            return;
+        }
+
         Schema::create('supply_orders', function (Blueprint $table) {
             $table->id();
             $table->foreignIdFor(\App\Models\PurchaseOrder::class)->constrained()->onDelete('cascade');
